only download/create files if they do not exist

+++ scripts/common.sh
@@ -49,10 +49,14 @@ sudo sysctl --system
 sudo apt-get update -y
 apt-get install -y software-properties-common curl apt-transport-https ca-certificates
 
-curl -fsSL https://pkgs.k8s.io/addons:/cri-o:/prerelease:/main/deb/Release.key |
+if [ ! -f "/etc/apt/keyrings/cri-o-apt-keyring.gpg" ]; then
+  curl -fsSL https://pkgs.k8s.io/addons:/cri-o:/prerelease:/main/deb/Release.key |
     gpg --dearmor -o /etc/apt/keyrings/cri-o-apt-keyring.gpg
+fi
+if [ ! -f "/etc/apt/sources.list.d/cri-o.list" ]; then
 echo "deb [signed-by=/etc/apt/keyrings/cri-o-apt-keyring.gpg] https://pkgs.k8s.io/addons:/cri-o:/prerelease:/main/deb/ /" |
     tee /etc/apt/sources.list.d/cri-o.list
+fi
 
 sudo apt-get update -y
 sudo apt-get install -y cri-o
@@ -64,8 +68,12 @@ sudo systemctl start crio.service
 echo "CRI runtime installed successfully"
 
 sudo mkdir -p /etc/apt/keyrings
-curl -fsSL https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/Release.key | sudo gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
-echo "deb [signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/ /" | sudo tee /etc/apt/sources.list.d/kubernetes.list
+if [ ! -f "/etc/apt/keyrings/kubernetes-apt-keyring.gpg" ]; then
+curl -fsSL https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/Release.key | gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
+fi
+if [ ! -f "/etc/apt/sources.list.d/kubernetes.list" ]; then
+echo "deb [signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/ /" | tee /etc/apt/sources.list.d/kubernetes.list
+fi
 
 
 sudo apt-get update -y
